Add enemy and camera effect
- add file to load and make enemy1
- add file for enemy facing left
- add function to make several enemies
- add camera effect to follow player
- add camera effect to only move up and down
  - and start at middle of screen
